fix sidenav behaviour in wire
avoid importing sidenav handling 2x

NHUB-617
