+++ assets/index.ts
@@ -1,12 +1,10 @@
 import {ComponentType} from 'react';
-import 'primereact/resources/primereact.min.css';
-
 import {IArticle, ICoverageMetadataPreviewProps} from 'interfaces';
-import {gettext, isTouchDevice} from 'utils';
 
-import {registerCoverageFieldComponent} from './agenda/components/preview';
+import {gettext} from 'utils';
 import {Button} from './ui/components/Button';
 
+import {registerCoverageFieldComponent} from './agenda/components/preview';
 
 interface IExtensions {
     prepareWirePreview?(content: HTMLElement, item: IArticle): HTMLElement;
@@ -46,37 +44,3 @@ export const exposed: IExposedForExtensions = {
 };
 
 import 'app';
-
-// navigation scripts
-if (isTouchDevice()) {
-    document.documentElement.classList.add('no-touch');
-}
-
-const navigation = document.getElementById('nav');
-const pinButton = document.getElementById('pin-btn');
-
-if (pinButton != null) {
-    pinButton.addEventListener('click', () => {
-        handleNavToggle();
-    });
-}
-
-if (sessionStorage.getItem('navigation-pinned')) {
-    handleNavToggle();
-}
-
-function handleNavToggle() {
-    if (navigation == null || pinButton == null) {
-        return null;
-    }
-
-    if (navigation.classList.contains('nav-block--pinned')) {
-        sessionStorage.removeItem('navigation-pinned');
-        navigation.classList.remove('nav-block--pinned');
-        pinButton.setAttribute('title', gettext('Expand'));
-    } else {
-        sessionStorage.setItem('navigation-pinned', 'true');
-        navigation.classList.add('nav-block--pinned');
-        pinButton.setAttribute('title', gettext('Collapse'));
-    }
-}
