+++ assets/style.ts
@@ -1,1 +1,2 @@
 import './styles/app.scss';
+import 'primereact/resources/primereact.min.css';
